+++ app/src/main/kotlin/com/maltaisn/notes/model/DefaultNotesRepository.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Nicolas Maltais
+ * Copyright 2023 Nicolas Maltais
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,13 +16,12 @@
 
 package com.maltaisn.notes.model
 
+import android.text.format.DateUtils
 import com.maltaisn.notes.model.entity.Note
 import com.maltaisn.notes.model.entity.NoteStatus
 import kotlinx.coroutines.NonCancellable
-import com.maltaisn.notes.ui.note.DeletedNotesTimeoutField
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
-import kotlin.time.Duration.Companion.days
 
 class DefaultNotesRepository @Inject constructor(
     private val notesDao: NotesDao,
@@ -74,12 +73,7 @@ class DefaultNotesRepository @Inject constructor(
     }
 
     override suspend fun deleteOldNotesInTrash() {
-        val delay = when (prefs.deletedNotesTimeoutField) {
-            DeletedNotesTimeoutField.DAY -> 1.days
-            DeletedNotesTimeoutField.DAYS_7 -> 7.days
-            DeletedNotesTimeoutField.MONTH -> 30.days
-            DeletedNotesTimeoutField.YEAR -> 365.days
-        }.inWholeMilliseconds
+        val delay = prefs.deletedNotesTimeout.value.toInt() * DateUtils.DAY_IN_MILLIS
         val minDate = System.currentTimeMillis() - delay
         notesDao.deleteNotesByStatusAndDate(NoteStatus.DELETED, minDate)
     }