+++ app/src/main/kotlin/com/maltaisn/notes/ui/note/DeletedNotesTimeoutField.kt
@@ -22,13 +22,13 @@ import com.maltaisn.notes.model.findValueEnum
 
 /**
  * Enum for deleted notes timeout.
- * [value] is from [R.array.pref_shown_date_values].
+ * [value] is from [R.array.pref_deleted_notes_timeout_values].
  */
 enum class DeletedNotesTimeoutField(override val value: String) : ValueEnum<String> {
-    DAY("24h"),
-    DAYS_7("7d"),
-    MONTH("30d"),
-    YEAR("1y");
+    DAY("1"),
+    WEEK("7"),
+    MONTH("30"),
+    YEAR("365");
 
     companion object {
         fun fromValue(value: String): DeletedNotesTimeoutField = findValueEnum(value)